Fixes exception when clicking Actions on empty pet detail screen. Actions button is now hidden when screen is opened with a null pet object, and the Actions button will check for a null state and terminate early.
Fixes exception when attempting to save an animal with status Missing.
The value of the status has been reverted back to "Lost" to match the name
of the status.

Fixes user can save a pet with no name bug. Save button now explicitly checks
that name field is not empty before calling saving method.